Allow for piping input, add pragmas for module code.
The first change is to optionally enable piping text
into glslify-stream instead of reading the file
directly. A second boolean argument is added to the
function such that:

    require('glslify-stream')(path, true)

Will not read the file and return a through stream
for writing to.

The second it to add pragmas to specify module start/end
points.

node_modules begin with:
    #pragma glslify_file start module
And other external files:
    #pragma glslify_file start
Both end with:
    #pragma glslify_file end

These two changes make it possible to add browserify-style
transform streams to glslify, in that they can be piped
in a chain and ignore module code.